refactor: 拆分巨石文件 - execution_request_service / similarity_service / jtn_case_import
拆分 3 个超大文件，提取子模块降低单文件复杂度：

1. execution_request_service.py (2063行 → 7个子模块)
   - execution_request_models.py: 6个dataclass
   - execution_request_utils.py: 共享工具函数和常量
   - execution_request_parser.py: 金额解析
   - execution_request_interest.py: 利息参数解析与计算
   - execution_request_clause_extractor.py: 条款提取(纯正则)
   - execution_request_llm_fallback.py: LLM/Ollama兜底解析
   - execution_request_text_generator.py: 文本生成

2. similarity_service.py (1362行 → 4个子模块)
   - similarity_scorers.py: 评分原语(classmethod)
   - similarity_passage.py: 段落选择/排序/去重
   - similarity_json_utils.py: JSON提取/修复/元数据
   - similarity_cache.py: 缓存读写/序列化

3. jtn_case_import.py (1979行 → 2个子模块)
   - jtn_case_import_models.py: 7个dataclass
   - jtn_case_html_parser.py: HTML解析纯函数

所有拆分保持外部导入路径不变(re-export)，302个单元测试全部通过。
