+++ backend/apps/oa_filing/services/oa_scripts/jtn_case_import.py
@@ -13,7 +13,6 @@ import os
 import re
 import time
 from concurrent.futures import ThreadPoolExecutor, as_completed
-from dataclasses import dataclass, field
 from typing import Any, Callable, Generator
 from urllib.parse import parse_qs, urljoin, urlparse
 
@@ -44,80 +43,17 @@ _MEDIUM_WAIT = 1.5
 _AJAX_WAIT = 2.0
 
 
-# ============================================================
-# 数据结构
-# ============================================================
-@dataclass
-class OACaseCustomerData:
-    """OA客户数据（案件中提取）。"""
-
-    name: str  # 客户名称
-    customer_type: str  # natural=自然人 / legal=企业
-    address: str | None = None  # 地址
-    phone: str | None = None  # 联系电话
-    id_number: str | None = None  # 身份证号码（自然人）
-    industry: str | None = None  # 行业（企业）
-    legal_representative: str | None = None  # 法定代表人（企业）
-
-
-@dataclass
-class OACaseInfoData:
-    """OA案件信息数据。"""
-
-    case_no: str  # 案件编号
-    case_name: str | None = None  # 案件名称
-    case_stage: str | None = None  # 案件阶段（一审/二审/执行）
-    acceptance_date: str | None = None  # 收案日期
-    case_category: str | None = None  # 案件类别/案件类型（合同类型映射主字段）
-    case_type: str | None = None  # 业务种类（兼容字段）
-    responsible_lawyer: str | None = None  # 案件负责人
-    description: str | None = None  # 案情简介
-    client_side: str | None = None  # 代理何方
-
-
-@dataclass
-class OAConflictData:
-    """OA利益冲突数据。"""
-
-    name: str  # 冲突方名称
-    conflict_type: str | None = None  # 冲突类型
-
-
-@dataclass
-class OACaseData:
-    """OA案件完整数据。"""
-
-    case_no: str  # 案件编号（OA案件编号）
-    keyid: str  # OA系统KeyID
-    customers: list[OACaseCustomerData] = field(default_factory=list)  # 客户列表
-    case_info: OACaseInfoData | None = None  # 案件信息
-    conflicts: list[OAConflictData] = field(default_factory=list)  # 利益冲突列表
-
-
-@dataclass
-class CaseSearchItem:
-    """案件搜索结果项。"""
-
-    case_no: str  # 案件编号
-    keyid: str  # 详情页KeyID
-
-
-@dataclass
-class OAListCaseCandidate:
-    """OA 列表页候选案件。"""
-
-    case_no: str
-    case_name: str
-    keyid: str
-    detail_url: str
-
-
-@dataclass
-class CaseListFormState:
-    """案件列表表单状态。"""
-
-    action_url: str
-    payload: dict[str, str]
+# 数据结构（从子模块导入）
+from .jtn_case_import_models import (
+    OAConflictData,
+    OACaseCustomerData,
+    OACaseData,
+    OACaseInfoData,
+    OAListCaseCandidate,
+    CaseListFormState,
+    CaseSearchItem,
+)
+from . import jtn_case_html_parser as html_parser
 
 
 # ============================================================
@@ -371,7 +307,7 @@ class JtnCaseImportScript:
 
         with httpx.Client(headers=_HTTP_HEADERS, follow_redirects=True, timeout=15, trust_env=False) as client:
             login_resp = client.get(_LOGIN_URL)
-            csrf_token = self._extract_hidden_input(login_resp.text, "CSRFToken")
+            csrf_token = html_parser.extract_hidden_input(login_resp.text, "CSRFToken")
 
             login_result = client.post(
                 _LOGIN_URL,
@@ -408,7 +344,7 @@ class JtnCaseImportScript:
         response.raise_for_status()
 
         next_form_state = self._extract_form_state(html_text=response.text, base_url=str(response.url), client=client)
-        keyid = self._extract_case_keyid_from_search_html(html_text=response.text, case_no=case_no)
+        keyid = html_parser.extract_case_keyid_from_search_html(html_text=response.text, case_no=case_no)
         if not keyid:
             return None, next_form_state
 
@@ -424,7 +360,7 @@ class JtnCaseImportScript:
         detail_url = _DETAIL_URL_TEMPLATE.format(base=_BASE_URL, keyid=search_item.keyid)
         response = client.get(detail_url)
         response.raise_for_status()
-        return self._parse_case_detail_html(
+        return html_parser.parse_case_detail_html(
             html_text=response.text,
             case_no=search_item.case_no,
             keyid=search_item.keyid,
@@ -450,7 +386,7 @@ class JtnCaseImportScript:
                 base_url=str(response.url),
                 client=client,
             )
-            candidates = self._extract_case_candidates_from_search_html(response.text)
+            candidates = html_parser.extract_case_candidates_from_search_html(response.text)
             return self._rank_name_candidates(keyword=keyword, candidates=candidates, limit=limit)
         except Exception:
             self._reset_name_search_http_session()
@@ -633,7 +569,7 @@ class JtnCaseImportScript:
             time.sleep(_SHORT_WAIT)
 
             html_text = target_frame.content()
-            candidates = self._extract_case_candidates_from_search_html(html_text)
+            candidates = html_parser.extract_case_candidates_from_search_html(html_text)
             return self._rank_name_candidates(keyword=keyword, candidates=candidates, limit=limit)
         except Exception as exc:
             if self._is_sso_blocking_error(exc):
@@ -677,11 +613,11 @@ class JtnCaseImportScript:
         candidates: list[OAListCaseCandidate],
         limit: int,
     ) -> list[OAListCaseCandidate]:
-        normalized = self._normalize_text(keyword)
+        normalized = html_parser.normalize_text(keyword)
         ordered_candidates = list(candidates)
         ordered_candidates.sort(
             key=lambda item: (
-                0 if normalized and normalized in self._normalize_text(item.case_name) else 1,
+                0 if normalized and normalized in html_parser.normalize_text(item.case_name) else 1,
                 item.case_no,
                 item.keyid,
             )
@@ -753,14 +689,14 @@ class JtnCaseImportScript:
                 continue
             option_value = option.get("value")
             payload[name] = (
-                str(option_value) if option_value is not None else self._normalize_text("".join(option.itertext()))
+                str(option_value) if option_value is not None else html_parser.normalize_text("".join(option.itertext()))
             )
 
         for textarea_node in form.xpath(".//textarea[@name]"):
             name = str(textarea_node.get("name") or "").strip()
             if not name:
                 continue
-            payload[name] = self._normalize_text("".join(textarea_node.itertext()))
+            payload[name] = html_parser.normalize_text("".join(textarea_node.itertext()))
 
         return CaseListFormState(action_url=action_url, payload=payload)
 
@@ -771,351 +707,6 @@ class JtnCaseImportScript:
         logger.warning("未找到指定案件名称查询字段: %s", _SEARCH_CASE_NAME_FIELD)
         return None
 
-    def _extract_case_candidates_from_search_html(self, html_text: str) -> list[OAListCaseCandidate]:
-        candidates: list[OAListCaseCandidate] = []
-        seen_keys: set[tuple[str, str]] = set()
-        try:
-            root = lxml_html.fromstring(html_text)
-            for row in root.xpath("//tr"):
-                links = row.xpath('.//a[contains(@href, "projectView.aspx") and contains(@href, "keyid=")]')
-                if not links:
-                    continue
-                cell_texts = self._extract_row_cells_text(row)
-                row_text = self._normalize_text(" ".join(cell_texts))
-                case_no = self._extract_case_no_from_text(row_text)
-
-                for link in links:
-                    href = str(link.get("href") or "")
-                    keyid = self._extract_keyid_from_href(href)
-                    if not keyid:
-                        continue
-
-                    case_name = self._normalize_text("".join(link.itertext()))
-                    if case_name in {"查看", "编辑", "删除", "详情", "操作"}:
-                        case_name = ""
-                    if not case_name:
-                        case_name = self._extract_case_name_from_row(
-                            row,
-                            case_no=case_no,
-                            row_text=row_text,
-                        )
-
-                    unique_key = (keyid, case_no)
-                    if unique_key in seen_keys:
-                        continue
-                    seen_keys.add(unique_key)
-
-                    candidates.append(
-                        OAListCaseCandidate(
-                            case_no=case_no,
-                            case_name=case_name,
-                            keyid=keyid,
-                            detail_url=_DETAIL_URL_TEMPLATE.format(base=_BASE_URL, keyid=keyid),
-                        )
-                    )
-        except Exception:
-            logger.debug("解析按名称查询结果失败", exc_info=True)
-
-        return candidates
-
-    def _extract_case_no_from_text(self, row_text: str) -> str:
-        text = self._normalize_text(row_text)
-        if not text:
-            return ""
-
-        patterns = [
-            r"(?<![A-Za-z0-9])\d{4}[A-Za-z]{1,8}\d{2,}(?![A-Za-z0-9])",
-            r"(?<![A-Za-z0-9])[A-Za-z]{1,4}\d{4,}(?![A-Za-z0-9])",
-            r"(?<![A-Za-z0-9])\d{6,}(?![A-Za-z0-9])",
-        ]
-        for pattern in patterns:
-            match = re.search(pattern, text)
-            if match:
-                return str(match.group(0)).strip()
-        return ""
-
-    def _extract_case_name_from_row(self, row_node: Any, *, case_no: str, row_text: str) -> str:
-        best_text = ""
-        best_score = -10_000
-
-        for cell_text in self._extract_row_cells_text(row_node):
-            score = self._score_case_name_cell(cell_text, case_no=case_no)
-            if score > best_score:
-                best_score = score
-                best_text = cell_text
-
-        cleaned = self._clean_case_name_text(best_text, case_no=case_no)
-        if cleaned:
-            return cleaned
-        return self._extract_case_name_from_row_text(row_text, case_no)
-
-    def _score_case_name_cell(self, cell_text: str, *, case_no: str) -> int:
-        text = self._normalize_text(cell_text)
-        if not text:
-            return -100
-        if text.isdigit():
-            return -90
-        if text in {"查看", "编辑", "删除", "详情", "操作"}:
-            return -80
-
-        score = 0
-        if case_no and case_no in text:
-            score += 30
-        if "诉" in text:
-            score += 20
-        if any(marker in text for marker in ("纠纷", "案件", "案【", "案[", "申请")):
-            score += 12
-        if any(marker in text for marker in ("[诉讼]", "民商事案件", "已完善", "信息完善", "在办中", "推送至社区")):
-            score -= 15
-        if re.fullmatch(r"\d{4}-\d{2}-\d{2}", text):
-            score -= 20
-        if len(text) <= 2:
-            score -= 10
-        return score
-
-    def _clean_case_name_text(self, value: str, *, case_no: str) -> str:
-        text = self._normalize_text(value)
-        if not text:
-            return ""
-
-        if case_no and case_no in text:
-            text = text.replace(case_no, " ")
-
-        for marker in ("查看", "编辑", "删除", "详情", "操作"):
-            text = text.replace(marker, " ")
-
-        for marker in (
-            "[诉讼]",
-            "[非诉]",
-            "民商事案件",
-            "刑事案件",
-            "行政案件",
-            "已完善",
-            "信息完善",
-            "在办中",
-            "修改承办律师",
-            "利冲变更申请",
-            "法院进程变更",
-            "保全信息变更",
-            "多地合作变更申请",
-            "对外合办变更申请",
-            "案件负责人变更申请",
-            "零收费变更申请",
-            "添加案件进程信息",
-            "添加工作日志审批人",
-            "上传定稿合同",
-            "取消推送至社区",
-            "推送至社区",
-            "已推业绩",
-            "撤销推送",
-        ):
-            if marker in text:
-                text = text.split(marker, 1)[0].strip()
-
-        text = re.sub(r"^(?:正\s*式|更多)\s+", "", text)
-        text = re.sub(r"^([\u4e00-\u9fa5A-Za-z·]{2,16})\s+(?=\1诉)", "", text)
-        text = re.sub(r"^\d+\s+", "", text)
-        return re.sub(r"\s+", " ", text).strip()
-
-    def _extract_case_name_from_row_text(self, row_text: str, case_no: str) -> str:
-        return self._clean_case_name_text(row_text, case_no=case_no)
-
-    def _extract_case_keyid_from_search_html(self, *, html_text: str, case_no: str) -> str | None:
-        """从查询结果 HTML 中解析案件 keyid。"""
-        try:
-            root = lxml_html.fromstring(html_text)
-            for row in root.xpath("//tr"):
-                row_text = self._normalize_text("".join(row.itertext()))
-                if case_no not in row_text:
-                    continue
-                links = row.xpath('.//a[contains(@href, "projectView.aspx") and contains(@href, "keyid=")]')
-                for link in links:
-                    href = str(link.get("href") or "")
-                    keyid = self._extract_keyid_from_href(href)
-                    if keyid:
-                        return keyid
-        except Exception:
-            logger.debug("lxml 解析查询结果失败，回退正则匹配: %s", case_no, exc_info=True)
-
-        escaped_case_no = re.escape(case_no)
-        regex = re.compile(
-            rf"{escaped_case_no}[\s\S]{{0,5000}}?projectView\.aspx\?keyid=([^&'\" >]+)",
-            re.IGNORECASE,
-        )
-        match = regex.search(html_text)
-        if match:
-            return match.group(1)
-        return None
-
-    def _extract_keyid_from_href(self, href: str) -> str | None:
-        if not href:
-            return None
-        full_url = urljoin(_CASE_LIST_URL, href)
-        query = parse_qs(urlparse(full_url).query)
-        keyid = query.get("keyid", query.get("KeyID", [None]))[0]
-        return str(keyid).strip() if keyid else None
-
-    def _parse_case_detail_html(
-        self,
-        *,
-        html_text: str,
-        case_no: str,
-        keyid: str,
-    ) -> OACaseData | None:
-        """解析案件详情 HTML（客户信息 + 案件信息 + 利冲）。"""
-        try:
-            root = lxml_html.fromstring(html_text)
-            customers = self._extract_customers_from_html(root)
-            case_info = self._extract_case_info_from_html(root, fallback_case_no=case_no)
-            conflicts = self._extract_conflicts_from_html(root)
-            return OACaseData(
-                case_no=case_no,
-                keyid=keyid,
-                customers=customers,
-                case_info=case_info,
-                conflicts=conflicts,
-            )
-        except Exception as exc:
-            logger.warning("解析案件详情HTML异常 %s: %s", case_no, exc)
-            return None
-
-    def _extract_customers_from_html(self, root: Any) -> list[OACaseCustomerData]:
-        customers: list[OACaseCustomerData] = []
-        rows = root.xpath('//div[@id="tab_con_1"]//tr')
-        current_customer: OACaseCustomerData | None = None
-
-        for row in rows:
-            cell_texts = self._extract_row_cells_text(row)
-            if not cell_texts:
-                continue
-
-            row_text = self._normalize_text(" ".join(cell_texts))
-            name_match = re.search(r"客户（([^）]+)）信息", row_text)
-            if name_match:
-                if current_customer and current_customer.name:
-                    customers.append(current_customer)
-
-                customer_name = self._normalize_text(name_match.group(1))
-                customer_type = "legal" if ("企业" in row_text or "公司" in customer_name) else "natural"
-                current_customer = OACaseCustomerData(name=customer_name, customer_type=customer_type)
-                continue
-
-            if current_customer is None:
-                continue
-
-            for label, value in self._iter_label_value_pairs(cell_texts):
-                if not label:
-                    continue
-                if "客户类型" in label and value:
-                    current_customer.customer_type = "legal" if "企业" in value or "公司" in value else "natural"
-                elif "身份证" in label and value:
-                    current_customer.id_number = value
-                elif "地址" in label and value:
-                    current_customer.address = value
-                elif ("法定代表" in label or "负责人" in label) and value:
-                    current_customer.legal_representative = value
-                elif "行业" in label and value:
-                    current_customer.industry = value
-                elif ("电话" in label or "号码" in label) and value:
-                    current_customer.phone = value
-
-        if current_customer and current_customer.name:
-            customers.append(current_customer)
-
-        return customers
-
-    def _extract_case_info_from_html(self, root: Any, *, fallback_case_no: str) -> OACaseInfoData:
-        case_info = OACaseInfoData(case_no=fallback_case_no)
-        rows = root.xpath('//div[@id="tab_con_2"]//tr')
-
-        for row in rows:
-            cell_texts = self._extract_row_cells_text(row)
-            if len(cell_texts) < 2:
-                continue
-
-            for label, value in self._iter_label_value_pairs(cell_texts):
-                if not label:
-                    continue
-                if "案件名称" in label and value:
-                    case_info.case_name = value
-                elif "案件阶段" in label and value:
-                    case_info.case_stage = value
-                elif "收案日期" in label and value:
-                    case_info.acceptance_date = value
-                elif ("案件类别" in label or "案件类型" in label) and value:
-                    case_info.case_category = value
-                elif "业务种类" in label and value:
-                    case_info.case_type = value
-                elif "案件负责人" in label and value:
-                    case_info.responsible_lawyer = value
-                elif "案情简介" in label and value:
-                    case_info.description = value[:500]
-                elif "代理何方" in label and value:
-                    case_info.client_side = value
-                elif "案件编号" in label and value:
-                    case_info.case_no = value
-
-        return case_info
-
-    def _extract_conflicts_from_html(self, root: Any) -> list[OAConflictData]:
-        conflicts: list[OAConflictData] = []
-        rows = root.xpath('//div[@id="tab_con_3"]//tr')
-
-        current_name: str | None = None
-        current_type: str | None = None
-
-        for row in rows:
-            cell_texts = self._extract_row_cells_text(row)
-            if not cell_texts:
-                continue
-
-            for label, value in self._iter_label_value_pairs(cell_texts):
-                if not label:
-                    continue
-                if "中文名称" in label and value:
-                    if current_name:
-                        conflicts.append(OAConflictData(name=current_name, conflict_type=current_type))
-                    current_name = value
-                    current_type = None
-                elif ("法律地位" in label and value) or (
-                    "类型" in label and "客户类型" not in label and "法律地位" not in label and value
-                ):
-                    current_type = value
-
-        if current_name:
-            conflicts.append(OAConflictData(name=current_name, conflict_type=current_type))
-
-        return conflicts
-
-    def _extract_row_cells_text(self, row_node: Any) -> list[str]:
-        cells = row_node.xpath("./td")
-        return [self._normalize_text("".join(cell.itertext())) for cell in cells]
-
-    def _iter_label_value_pairs(self, cell_texts: list[str]) -> list[tuple[str, str]]:
-        pairs: list[tuple[str, str]] = []
-        for idx in range(0, len(cell_texts) - 1, 2):
-            label = self._normalize_label(cell_texts[idx])
-            value = self._normalize_text(cell_texts[idx + 1])
-            pairs.append((label, value))
-        return pairs
-
-    def _normalize_text(self, value: str | None) -> str:
-        if value is None:
-            return ""
-        compact = str(value).replace("\xa0", " ").replace("　", " ")
-        return re.sub(r"\s+", " ", compact).strip()
-
-    def _normalize_label(self, value: str | None) -> str:
-        text = self._normalize_text(value)
-        return text.replace("：", "").replace(":", "").replace(" ", "")
-
-    def _extract_hidden_input(self, html_text: str, name: str) -> str:
-        pattern = re.compile(
-            rf'<input[^>]+name=["\']{re.escape(name)}["\'][^>]*value=["\']([^"\']*)["\']',
-            re.IGNORECASE,
-        )
-        match = pattern.search(html_text)
-        return match.group(1).strip() if match else ""
 
     def _is_login_failed_response(self, response: httpx.Response) -> bool:
         """根据登录响应判断是否仍停留在登录失败状态。"""
